+++ fork/dynamic-cdn-webpack-plugin/test/helpers/clean-dir.js
@@ -1,14 +1,6 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
-const rimraf = require('rimraf');
+const { rimraf } = require('rimraf');
 
 module.exports = function cleanDir(dir) {
-	return new Promise((resolve, reject) => {
-		rimraf(dir, err => {
-			if (err) {
-				return reject(err);
-			}
-
-			return resolve();
-		});
-	});
+	return rimraf(dir);
 };