+++ fork/json-schema-form-core/webpack.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+	output: {
+		filename: 'index.js',
+	},
+	resolve: {
+		fallback: {
+			path: require.resolve('path-browserify'),
+		},
+	},
+};