+++ packages/components/src/QualityBar/QualityBar.stories.tsx
@@ -1,4 +1,5 @@
 import { action } from '@storybook/addon-actions';
+
 import { QualityBar } from './QualityBar.component';
 
 export default {