+++ packages/forms/src/UIForm/fields/FieldTemplate/displayMode/TextMode.component.test.js
@@ -1,19 +1,20 @@
-import { screen, render } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 
 import TextModeFieldTemplate from './TextMode.component';
 
+jest.unmock('@talend/design-system');
 describe('FieldTemplate in text display mode', () => {
-	it('should render dt/dd', () => {
+	it('should render text', () => {
 		// when
 		const { container } = render(
 			<TextModeFieldTemplate id="myAwesomeField" label="My awesome label">
-				My value as chrildren
+				My value as children
 			</TextModeFieldTemplate>,
 		);
 
 		// then
-		expect(screen.getByRole('term')).toHaveTextContent('My awesome label');
-		expect(screen.getByRole('definition')).toHaveTextContent('My value as chrildren');
+		expect(screen.getByText('My awesome label')).toBeInTheDocument();
+		expect(screen.getByText('My value as children')).toBeInTheDocument();
 		expect(container.firstChild).toMatchSnapshot();
 	});
 });