+++ packages/forms/src/index.ts
@@ -1,10 +1,9 @@
+import { I18N_DOMAIN_FORMS } from './constants';
 import FormSkeleton from './FormSkeleton';
 import FormSwitcher from './FormSwitcher';
-import UIForm, { Code } from './UIForm';
-import { I18N_DOMAIN_FORMS } from './constants';
 import rhf from './rhf';
+import UIForm, { Code } from './UIForm';
 import validate from './validate';
-import widgets from './widgets';
 
 const constants = {
 	I18N_DOMAIN_FORMS,
@@ -14,7 +13,7 @@ const utils = {
 	validate,
 };
 
-export { FormSwitcher, constants, utils, FormSkeleton, rhf, UIForm, widgets };
+export { FormSwitcher, constants, utils, FormSkeleton, rhf, UIForm };
 
 export type { DisplayMode, UiSchema, FormDefinition } from './types/index';
 
@@ -24,6 +23,5 @@ export default Object.assign(FormSwitcher, {
 	Skeleton: FormSkeleton,
 	rhf,
 	UIForm,
-	widgets,
 	Code,
 });