+++ packages/forms/src/widgets/templates/FieldTemplate/index.js
@@ -1,3 +1,0 @@
-import FieldTemplate from './FieldTemplate.component';
-
-export default FieldTemplate;