+++ packages/playground/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @talend/ui-playground
 
+## 2.4.4
+
+### Patch Changes
+
+- Updated dependencies [27a4a5d]
+  - @talend/react-faceted-search@19.0.0
+
+## 2.4.3
+
+### Patch Changes
+
+- Updated dependencies [0629df7]
+  - @talend/react-forms@14.0.0
+  - @talend/react-containers@10.7.3
+
 ## 2.4.2
 
 ### Patch Changes