+++ tools/scripts-publish-local/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @talend/scripts-publish-local
 
+## 1.3.3
+
+### Patch Changes
+
+- af0ac2d: Upgrade rimraf to version 5.0.5
+
 ## 1.3.2
 
 ### Patch Changes